fix(arm64): e_effectlist — pointer truncation in SMP thread param
The SMP thread callback receives a void* that encodes an integer index.
On 64-bit platforms (LP64) casting a pointer directly to int drops the
upper 32 bits; cast via intptr_t first to silence the truncation and
avoid UB on platforms where the pointer value exceeds INT_MAX.

+++ avs/vis_avs/e_effectlist.cpp
@@ -627,7 +627,7 @@ void E_EffectList::smp_render_list(int min_threads,
 }
 
 uint32_t E_EffectList::smp_thread_proc(void* parm) {
-    int which = (int)parm;
+    int which = (int)(intptr_t)parm;
     void* hdls[2] = {E_EffectList::smp.thread_start[which],
                      E_EffectList::smp.quit_signal};
     for (;;) {
